Refactor Logic and storage components to use Contact model
- Updated Logic interface and LogicManager to return ObservableList<Contact> instead of ObservableList<Person>.
- Changed imports in JsonAdaptedTrip and its test to reflect the use of Contact model instead of Person.
