+++ src/main/java/seedu/address/logic/Logic.java
@@ -8,8 +8,6 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyTripBook;
-import seedu.address.model.trip.Trip;
 import seedu.address.model.contact.Contact;
 
 /**
@@ -33,7 +31,7 @@ public interface Logic {
     ReadOnlyAddressBook getAddressBook();
 
     /** Returns an unmodifiable view of the filtered list of persons */
-    ObservableList<Person> getFilteredPersonList();
+    ObservableList<Contact> getFilteredPersonList();
 
     /**
      * Returns the user prefs' address book file path.
