+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -15,8 +15,6 @@ import seedu.address.logic.parser.AddressBookParser;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyTripBook;
-import seedu.address.model.trip.Trip;
 import seedu.address.model.contact.Contact;
 import seedu.address.storage.Storage;
 
@@ -69,7 +67,7 @@ public class LogicManager implements Logic {
     }
 
     @Override
-    public ObservableList<Person> getFilteredPersonList() {
+    public ObservableList<Contact> getFilteredPersonList() {
         return model.getFilteredPersonList();
     }
